Add fitur
berhasil membuat search
berhasil membuat user
berhasil menghapus user
berhasil menampilkan detail

Pr update user apakah sudah bisa ?

+++ app/Http/Controllers/Component/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Component;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Contracts\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 
@@ -12,33 +13,21 @@ use App\Models\biodata_user as Biodata;
 
 class UserController extends Controller
 {
-    public function create(Request $request){
-        //  ini validate inline
-        $value = $request->validate([
-            'email' => ['required','email','unique:users,email'],
-            'password' => ['required','string','min:8'],
-            'nama_lengkap' => ['required','string'],
-            'no_handphone' => ['required','string'],
-            'alamat' => ['nullable'],
-        ]);
-        // craete
+    public function create(String $email,String $password , String $nama_lengkap , String $no_hp ,$alamat = null){
         $user = User::create([
-            'email' => $value['email'],
-            'password' => Hash::make($value['password']),
+            'email' => $email,
+            'password' => Hash::make($password),
         ]);
 
         $biodata = Biodata::create([
             'user_id' => $user->id,
-            'nama_lengkap' => $value['nama_lengkap'],
-            'no_handphone' => $value['no_handphone'],
-            'alamat' => (isset($value['alamat'])) ? $value['alamat'] : "belum di setting",
+            'nama_lengkap' => $nama_lengkap,
+            'no_handphone' => $no_hp,
+            'alamat' => (isset($alamat)) ? $alamat : "belum di setting",
         ]);
+        $user->assignRole('admin');
 
-        return response()->json([
-            'status' => [
-                'pesan' => "Berhasil membuat user",
-            ],
-        ],200);
+        return true;
     }
 
     public function update(Request $request){
@@ -51,15 +40,15 @@ class UserController extends Controller
             'alamat' => ['nullable'],
         ]);
         $biodata = Biodata::find($values['id_user']);
-        $biodata['email'] = $biodata->user->email;
+        $biodata['email'] = $biodata->User->email;
        
         if(isset($values['email'])){
-            $biodata->user->update([
+            $biodata->User->update([
                 'email' => $values['email'],
             ]);
         }
         if(isset($values['password'])){
-            $biodata->user->update([
+            $biodata->User->update([
                 'password' => Hash::make($values['password']),
             ]);
         }
@@ -86,6 +75,22 @@ class UserController extends Controller
         ],200);
     }
 
+    public function search(String $search = null , int $page_number = 2){
+        // pencarian 
+        if(empty($search)){
+            $biodata = Biodata::paginate($page_number);
+            return $biodata;
+        }
+        $biodata = Biodata::whereHas('User',function(Builder $q) use ($search) {
+            $q->where('email','like','%'.$search.'%');
+        })
+        ->orWhere('nama_lengkap','like','%'.$search.'%')
+        ->orWhere('no_handphone','like','%'.$search.'%')
+        ->paginate($page_number);
+
+        return $biodata;
+    }
+
 
     public function detail($id_user = null){
         if(empty($id_user)) return null;
@@ -94,12 +99,12 @@ class UserController extends Controller
         return $biodata;
     }
 
-    public function delete(Request $request){
-        $value = $request->validate([
-            'id_user' => ['required'],
-        ]);
-        $biodata = Biodata::find($value['id_user']);
-        $biodata->user->delete();
+    public function delete($id_user = null){
+        if(empty($id_user)) return null;
+        $biodata = Biodata::find($id_user);
+        $user = User::find($biodata['user_id']);
         $biodata->delete();
+        $user->delete();
+        return true;
     }
 }
